+++ tests/test_pvalues.cpp
@@ -22,15 +22,15 @@ TEST_CASE("p-value integration test") {
     // Load data used in calculation
     OtuTable otu_table;
     otu_table.load_otu_file(otu_fp);
-    arma::Mat<double> observed_correlation = load_correlation_file(observed_correlation_fp);
+    arma::Mat<float> observed_correlation = load_correlation_file(observed_correlation_fp);
 
     // Calculate pvalues
-    arma::Mat<double> test_pseudo_pvalues = calculate_pvalues(otu_table, observed_correlation, bs_correlation_fps, 3, false, 1);
-    arma::Mat<double> test_exact_pvalues = calculate_pvalues(otu_table, observed_correlation, bs_correlation_fps, 3, true, 1);
+    arma::Mat<float> test_pseudo_pvalues = calculate_pvalues(otu_table, observed_correlation, bs_correlation_fps, 3, false, 1);
+    arma::Mat<float> test_exact_pvalues = calculate_pvalues(otu_table, observed_correlation, bs_correlation_fps, 3, true, 1);
 
     // Load data for comparison (pseudo p-values from SparCC, exact previously validated)
-    arma::Mat<double> pseudo_pvalues = load_correlation_file(pseudo_fp);
-    arma::Mat<double> exact_pvalues = load_correlation_file(exact_fp);
+    arma::Mat<float> pseudo_pvalues = load_correlation_file(pseudo_fp);
+    arma::Mat<float> exact_pvalues = load_correlation_file(exact_fp);
 
     // Tolerating difference of 0.0001 as output types are rounded
     REQUIRE(arma::approx_equal(test_pseudo_pvalues, pseudo_pvalues, "absdiff", 0.001));
@@ -50,7 +50,7 @@ TEST_CASE("factorial division") {
     //  required is cancelation of factorials during division to avoid overflow
     // Data
     int sample_number = 32;
-    std::unordered_map<double, int> count_frequency = {{10, 1}, {0, 20}, {2, 5}, {3, 6}};
-    double result = calculate_possbile_otu_permutations(count_frequency, sample_number);
+    std::unordered_map<float, int> count_frequency = {{10, 1}, {0, 20}, {2, 5}, {3, 6}};
+    float result = calculate_possbile_otu_permutations(count_frequency, sample_number);
     REQUIRE(1251795504960 == result);
 }
